Move all file operations to async task
Move all file operations to async task
Change cache to limited size LruCache

+++ FeedMagic/CustomChoiceList/src/main/java/com/example/android/feedmagic/DownloadImagesTask.java
@@ -3,14 +3,18 @@ package com.example.android.feedmagic;
 import android.graphics.Bitmap;
 import android.graphics.BitmapFactory;
 import android.os.AsyncTask;
+import android.util.Log;
+import android.util.LruCache;
 import android.widget.ImageView;
 
 import java.io.File;
+import java.io.FileInputStream;
 import java.io.FileOutputStream;
 import java.io.IOException;
 import java.io.InputStream;
 import java.net.MalformedURLException;
 import java.net.URL;
+import java.util.HashMap;
 
 /**
  * Created by pdamu on 2/11/14.
@@ -18,16 +22,23 @@ import java.net.URL;
 public class DownloadImagesTask extends AsyncTask<ImageView, Void, Bitmap> {
 
     ImageView imageView = null;
+    LruCache<String, Bitmap> mBitmapHashMap;
+
+    public DownloadImagesTask(LruCache<String, Bitmap> bitmapHashMap) {
+        super();
+        mBitmapHashMap = bitmapHashMap;
+    }
 
     @Override
     protected Bitmap doInBackground(ImageView... imageViews) {
         this.imageView = imageViews[0];
-        return download_Image((String)imageView.getTag());
+        return download_Image((String) imageView.getTag());
     }
 
     @Override
     protected void onPostExecute(Bitmap result) {
-        if(result != null) {
+        if (result != null) {
+            mBitmapHashMap.put((String)imageView.getTag(), result);
             imageView.setImageBitmap(result);
         }
     }
@@ -36,24 +47,32 @@ public class DownloadImagesTask extends AsyncTask<ImageView, Void, Bitmap> {
     private Bitmap download_Image(String url) {
         Bitmap bitmap = null;
         try {
-            bitmap = BitmapFactory.decodeStream((InputStream) new URL(url).getContent());
-            try {
-                File file = new File("data/data/com.example.android.feedmagic/feedimg-" + url.hashCode());
-                if(!file.exists()){
-                    file.createNewFile();
+            File file = new File("data/data/com.example.android.feedmagic/feedimg-" + url.hashCode());
+            if (file.exists()) {
+                try {
+                    bitmap = BitmapFactory.decodeStream(new FileInputStream(file));
+                    Log.v("feedmagic","Reading from file..  " + url);
+                } catch (Exception e) {
+                    e.printStackTrace();
                 }
+            } else {
                 FileOutputStream out = new FileOutputStream(file);
-                bitmap.compress(Bitmap.CompressFormat.PNG, 90, out);
-                out.flush();
-                out.close();
-            } catch (Exception e) {
-                e.printStackTrace();
+                try {
+                    bitmap = BitmapFactory.decodeStream((InputStream) new URL(url).getContent());
+                    if (!file.exists()) {
+                        file.createNewFile();
+                    }
+                    bitmap.compress(Bitmap.CompressFormat.PNG, 90, out);
+                } catch (Exception e) {
+                    e.printStackTrace();
+                } finally {
+                    out.flush();
+                    out.close();
+                }
             }
-        } catch (MalformedURLException e) {
-            e.printStackTrace();
-        } catch (IOException e) {
+        } catch (Exception e) {
             e.printStackTrace();
         }
         return bitmap;
     }
-    }
+}
